fix(#198): make the canary's restore honest, and run its self-tests without FastAPI
Three findings from the PR #210 review:

[1] A `missing` restore reported success even when the delete never happened.
`restore` returned True unconditionally for `missing`, and `_remove` swallows
its own failures — so a leaked file whose parent went unreadable (0o000) mid-run
was left on disk while the report said "restored to its pre-run state". Re-check
that the path is provably gone (`snapshot(path) == MISSING`) and report
"COULD NOT BE RESTORED" otherwise.

[2] "restored to its pre-run state" over-claimed. `Entry.fingerprint` carries
`st_mtime_ns`, but `restore` put back only bytes and mode, so `snapshot(path)`
!= `before` the instant after a "successful" restore. Adopt option (a): stamp
the mtime back with `os.utime(ns=...)` (`follow_symlinks=False` for symlinks,
skipped where the platform lacks lutimes — the one documented narrowing to
bytes-and-mode). The claim is now literally true.

[3] The pure `env_sandbox` helper tests sat behind `test_env_isolation.py`'s
`pytest.importorskip("fastapi")`, so a machine without FastAPI skipped the
harness's own self-tests wholesale. Split them into the new
`tests/test_env_sandbox_helpers.py` (no FastAPI, no sandbox fixtures) and narrow
the gate in `test_env_isolation.py` to the single test that drives the web stack.

Moved tests keep their names verbatim, so the existing mutation table maps 1:1
onto the new module (snapshot/restore/leak_report/_listing/_park/Entry unit
tests — every `test_canary_*`, `test_snapshot_*`, `test_restore_*`,
`test_the_directory_*`, `test_the_listing_*`, `test_a_*`, `test_leak_report_*`,
`test_the_rescue_copy_*`). Left in `test_env_isolation.py`: the fixture/seal
integration tests (`test_pytest_cannot_reach_the_real_app_config`,
`test_settings_switch_*` [FastAPI-gated], `test_active_root_fallback_*`,
`test_module_scoped_*`, `test_the_session_hook_survives_a_canary_that_raises`,
`test_the_real_config_baseline_is_captured_at_import_time`).

New regression tests (in the helper module):
- test_canary_reports_a_missing_restore_it_could_not_carry_out
  mutation: revert the re-check to `return True` -> red (message reads
  "restored to its pre-run state", not "COULD NOT BE RESTORED").
- test_restore_returns_the_path_to_its_exact_pre_run_state_mtime_and_all
- test_restore_returns_a_symlink_to_its_exact_pre_run_state
  mutation: drop the `_restore_mtime` calls -> red (`snapshot(path) != before`;
  fingerprint mtime_ns is "now", not the pinned pre-run value).
